+++ package.json
@@ -2,17 +2,18 @@
   "name": "element-ui-el-table-draggable",
   "version": "1.1.6",
   "description": "Make 'el-table' of 'element UI' sortable.",
-  "main": "dist/SortableElTable.umd.js",
-  "module": "dist/SortableElTable.esm.js",
-  "unpkg": "dist/SortableElTable.min.js",
-  "browser": {
-    "./sfc": "src/SortableElTable.vue"
-  },
+  "main": "dist/element-ui-el-table-draggable.ssr.js",
+  "module": "dist/element-ui-el-table-draggable.esm.js",
+  "unpkg": "dist/element-ui-el-table-draggable.min.js",
+  "files": [
+    "dist/*",
+    "src/**/*.vue"
+  ],
   "scripts": {
-    "build": "npm run build:umd & npm run build:es & npm run build:unpkg",
-    "build:umd": "rollup --config build/rollup.config.js --format umd --file dist/SortableElTable.umd.js",
-    "build:es": "rollup --config build/rollup.config.js --format es --file dist/SortableElTable.esm.js",
-    "build:unpkg": "rollup --config build/rollup.config.js --format iife --file dist/SortableElTable.min.js"
+    "build": "cross-env NODE_ENV=production rollup --config build/rollup.config.js",
+    "build:ssr": "cross-env NODE_ENV=production rollup --config build/rollup.config.js --format cjs",
+    "build:es": "cross-env NODE_ENV=production rollup --config build/rollup.config.js --format es",
+    "build:unpkg": "cross-env NODE_ENV=production rollup --config build/rollup.config.js --format iife"
   },
   "repository": {
     "type": "git",
@@ -28,16 +29,19 @@
     "url": "https://github.com/WakuwakuP/element-ui-el-table-draggable/issues"
   },
   "homepage": "https://github.com/WakuwakuP/element-ui-el-table-draggable#readme",
-  "devDependencies": {
-    "rollup": "^0.57.1",
-    "rollup-plugin-buble": "^0.19.2",
-    "rollup-plugin-vue": "^3.0.0",
-    "tar": "^4.4.8",
-    "vue": "^2.6.10",
-    "vue-template-compiler": "^2.6.10",
-    "vue-template-validator": "^1.1.5"
-  },
   "dependencies": {
     "sortablejs": "^1.9.0"
+  },
+  "devDependencies": {
+    "cross-env": "^5.2.0",
+    "minimist": "^1.2.0",
+    "rollup": "^1.12.1",
+    "rollup-plugin-buble": "^0.19.6",
+    "rollup-plugin-commonjs": "^10.0.0",
+    "rollup-plugin-replace": "^2.2.0",
+    "rollup-plugin-terser": "^4.0.4",
+    "rollup-plugin-vue": "^5.0.0",
+    "vue": "^2.6.10",
+    "vue-template-compiler": "^2.6.10"
   }
 }
